requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
attrs==25.1.0
base58==2.1.1
bitarray==3.1.1
bitcoinlib==0.7.2
cbor2==5.6.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.1.0
coincurve==20.0.0
colorama==0.4.6
Colorthon==2.9.6
crcmod==1.7
CryptoFuzz==18.6.0
Cython==3.0.12
cytoolz==1.0.1
ecdsa==0.19.1
ed25519-blake2b==1.4.1
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastecdsa==3.0.1
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hdwallet==3.3.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
linkook==1.1.0
meson==1.7.0
meson-python==0.17.1
mnemonic==0.21
multidict==6.1.0
numpy==2.2.3
packaging==24.2
parsimonious==0.10.0
propcache==0.3.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
pyproject-metadata==0.9.1
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
requests_random_user_agent==2023.10.25
rlp==4.1.0
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
tronpy==0.5.0
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
versioneer==0.29
web3==7.9.0
websockets==13.1
wheel==0.45.1
yarl==1.18.3